Partial review of text and code
